+++ src/SignalR/server/Core/src/HubInvocationContext.cs
@@ -20,9 +20,9 @@ namespace Microsoft.AspNetCore.SignalR
         /// Instantiates a new instance of the <see cref="HubInvocationContext"/> class.
         /// </summary>
         /// <param name="context">Context for the active Hub connection and caller.</param>
-        /// <param name="serviceProvider"></param>
+        /// <param name="serviceProvider">The <see cref="IServiceProvider"/> specific to the scope of this Hub method invocation.</param>
         /// <param name="hub">The instance of the Hub.</param>
-        /// <param name="hubMethod"></param>
+        /// <param name="hubMethod">The <see cref="MethodInfo"/> for the Hub method being invoked.</param>
         /// <param name="hubMethodArguments">The arguments provided by the client.</param>
         public HubInvocationContext(HubCallerContext context, IServiceProvider serviceProvider, Hub hub, MethodInfo hubMethod, object[] hubMethodArguments)
 #pragma warning disable CS0618 // Type or member is obsolete
@@ -77,8 +77,14 @@ namespace Microsoft.AspNetCore.SignalR
         /// </summary>
         public IReadOnlyList<object> HubMethodArguments { get => Arguments; }
 
+        /// <summary>
+        /// The <see cref="IServiceProvider"/> specific to the scope of this Hub method invocation.
+        /// </summary>
         public IServiceProvider ServiceProvider { get; }
 
+        /// <summary>
+        /// The <see cref="MethodInfo"/> for the Hub method being invoked.
+        /// </summary>
         public MethodInfo HubMethod { get; }
     }
 }
