+++ src/SignalR/server/Core/src/HubLifetimeContext.cs
@@ -5,9 +5,17 @@ using System;
 
 namespace Microsoft.AspNetCore.SignalR
 {
-    // TODO: naming
+    /// <summary>
+    /// Context for the hub lifetime events <see cref="Hub.OnConnectedAsync"/> and <see cref="Hub.OnDisconnectedAsync(Exception)"/>.
+    /// </summary>
     public class HubLifetimeContext
     {
+        /// <summary>
+        /// Instantiates a new instance of the <see cref="HubLifetimeContext"/> class.
+        /// </summary>
+        /// <param name="context">Context for the active Hub connection and caller.</param>
+        /// <param name="serviceProvider">The <see cref="IServiceProvider"/> specific to the scope of this Hub method invocation.</param>
+        /// <param name="hub">The instance of the Hub.</param>
         public HubLifetimeContext(HubCallerContext context, IServiceProvider serviceProvider, Hub hub)
         {
             Hub = hub;
@@ -25,6 +33,9 @@ namespace Microsoft.AspNetCore.SignalR
         /// </summary>
         public Hub Hub { get; }
 
+        /// <summary>
+        /// The <see cref="IServiceProvider"/> specific to the scope of this Hub method invocation.
+        /// </summary>
         public IServiceProvider ServiceProvider { get; }
     }
 }
